refactoring
Major refcatoring to meet homework guidelines.
Need to add a seed for the ROI pixel but for now I will finish the other exercise so that the assignment can be reasonably finished before the due date.
I've changed the functions to only use the ones allowed on the asignment i.e. plotly, numpy, seaborn, matplotlib and pillow. I need to make sure the potly movies work on colab because I'm not sure how to implement this function yet in colab. '